+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/trackselection/TrackSelection.java
@@ -31,8 +31,8 @@ import java.lang.annotation.Target;
  * A track selection consisting of a static subset of selected tracks belonging to a {@link
  * TrackGroup}.
  *
- * <p>Tracks belonging to the subset are exposed in selection order, which by default is decreasing
- * bandwidth order.
+ * <p>Tracks belonging to the subset are exposed in selection order with the highest priority track
+ * first, for example in decreasing bandwidth order.
  */
 @UnstableApi
 public interface TrackSelection {